requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.28.0
psycopg2-binary==2.9.9
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
python-dotenv==1.0.0
pydantic-settings==2.0.3
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.1
tenacity==8.2.3
sqlalchemy-utils==0.41.1
greenlet==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
